Flattened AbstractWalWriterFuzzTest and AbstractFuzzTest
