+++ core/src/test/java/io/questdb/test/griffin/FrameAppendFuzzTest.java
@@ -130,7 +130,8 @@ public class FrameAppendFuzzTest extends AbstractFuzzTest {
         ff.close(metaFd);
     }
 
-    private void runFuzz(Rnd rnd) throws Exception {
+    @Override
+    protected void runFuzz(Rnd rnd) throws Exception {
         configOverrideO3ColumnMemorySize(rnd.nextInt(16 * 1024 * 1024));
 
         assertMemoryLeak(() -> {
